tests: drivers: gpio: gpio_basic_api: Add support for kv260
Add gpio driver testcase support for kv260. The overlay will select the
PS gpio bank to be tested.
Add test scenario name with filter and etra_arg for kv260.

For running this testcase on kv260 PS gpio bank 3 has been chosen
because it is EMIO and design is required to connect pins together.

Signed-off-by: Paul Alvin <alvin.paulp@amd.com>
Message-ID: <20250214060343.13695-3-alvin.paulp@amd.com>
Link: https://github.com/zephyrproject-rtos/zephyr/pull/85798/
State: waiting