boards: amd: versalnet: Wire AMD QEMU via west and twister
Wire AMD QEMU via qemu-system-xilinx-aarch64 in SDK
The latest AMD QEMU version is not available in any official SDK.
But you can build it yourself and add it to existing Zephyr SDK.

Download the latest QEMU:
git clone https://github.com/Xilinx/qemu

Configure it (align your SDK location):
mkdir test && cd test && \
../qemu/configure --disable-docs --disable-sdl --disable-debug-info \
--disable-cap-ng --disable-libnfs --disable-libusb --disable-libiscsi \
--disable-usb-redir --disable-linux-aio --disable-guest-agent \
--disable-libssh --disable-seccomp --disable-tpm --disable-numa \
--disable-glusterfs --disable-virtfs --disable-xen --disable-curl \
--disable-attr --disable-curses --disable-iconv --disable-kvm \
--disable-parallels --disable-replication --disable-live-block-migration \
--target-list="aarch64-softmmu" --skip-meson \
--enable-gcrypt \
--prefix=XYZ/zephyr-sdk/zephyr-sdk-XYZ/sysroots/x86_64-pokysdk-linux/usr/xilinx/

There is an attempt to enable this configuration in official SDK
https://github.com/zephyrproject-rtos/sdk-ng/pull/859
Difference in flags is calling with additional "--enable-gcrypt" to
properly model cryptographics devices.
Zephyr is not using them but it is easier for keeping configuration in sync
with official DTSes.

The patch is adding versalnet_rpu-qemu.dts which is export from DTB got via
https://github.com/Xilinx/qemu-devicetrees
(file LATEST/SINGLE_ARCH/board-versal-net-psx-vn-p-b2197-00-x-prc-09.dtb)
where only one change has been made in serial@0xf1920000
- chardev = "serial2";
+ chardev = "con";

The reason for this change is that Zephyr scripts are referencing to
chardev with name "con" as console.
"-chardev stdio,id=con,mux=on -serial chardev:con -mon chardev=con,mode=readline"
This is not a problem when you have system with only one serial but default
AMD DTSes are describing multiple consoles that's why it is necessary to
properly point to serial IP which is used as output device.

The second important change is adding CMakeLists.txt which is describing
building DTB out of ${BOARD}-qemu.dts to have full QEMU description in the
tree in text format instead of dealing with binaries as is done for QEMU
boards (./boards/qemu/cortex_a9/fdt-zynq7000s.dtb,
./boards/qemu/cortex_r5/fdt-single_arch-zcu102-arm.dtb).
The biggest advantage is that small changes (like chardev above) can be
easier review in text form and also anybody can do changes there.

DT description is terrible because it is DTB export and a lot of things
should be fixed (for example: node names) but there is no tool check
against dt binding because there is no binding for QEMU models yet. That's
why dtc is called with -q to suppress all the warnings coming from dts->dtb
conversion.

The main future is that west build -t run_qemu is wired which improves way
how to test this system.

Signed-off-by: Michal Simek <michal.simek@amd.com>
Message-ID: <ee3633ef012aaf09bd50d07610cf738c28f9df28.1739959112.git.michal.simek@amd.com>
State: pending