tests: drivers: watchdog: wdt_basic_api: Add support for Xilinx WWDT
Set maximum window and disable callbacks to verify that the watchdog
timer install/setup/feed functions correctly and that a reset is
triggered upon timeout.

The test_wdt_bad_window_max is ignored for Xilinx WWDT because the
driver uses the value of DT timeout-sec property for window.max when
window.max is set to zero from userspace.

Signed-off-by: Harini T <harini.t@amd.com>
Message-ID: <20250226040246.1041137-5-harini.t@amd.com>
State: pending