tests: drivers: watchdog: wdt_error_cases: Add support for Xilinx WWDT
Add configuration details for Xilinx WWDT.

Modify test_04w_wdt_install_timeout_with_invalid_window to verify
timeout installation in invalid window.

Add test_08f_wdt_disable_in_second_window to verify disabling in an
invalid window.

The testcases that involve using different watchdog options are not
supported because the driver does not support options for pausing in
sleep or when halted by the debugger.
The testcases with callbacks are also not supported.

Skip the tests that are not relevant for Xilinx WWDT.

Signed-off-by: Harini T <harini.t@amd.com>
Message-ID: <20250226040246.1041137-6-harini.t@amd.com>
State: pending