Revert "drivers: spi_xlnx_axi_quadspi: Reduce IRQ work"
This reverts commit 680fa154bf7896e2a0aa8d973d0422097c6d13aa.

'commit 680fa154bf78 ("drivers: spi_xlnx_axi_quadspi: Reduce IRQ work")'
introduces the use of events in non-async SPI transfers to signal the
calling thread upon completion. However, in the case of small transfers,
it has been observed that the event is posted even before the calling
thread starts waiting for it. As a result, the calling thread never
receives the event, causing the transfer to fail with a DTR empty
timeout error. This issue was identified while performing flash
tests 'tests/drivers/flash/common/' with this driver.

Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@amd.com>
Message-ID: <20250310100115.33377-3-amitrkcian2002@gmail.com>
Link: https://github.com/zephyrproject-rtos/zephyr/issues/86799
Link: https://github.com/zephyrproject-rtos/zephyr/pull/86809
State: waiting