dts: bindings: dma: Add Xilinx ADMA driver
Xilinx ADMA in PS can be used as general purpose DMA. It is designed to
support memory to memory transfer and memory to IO buffer transfer.
There are 8/16 instances of this DMA dependent on the device family. Add
support for Versal Gen 2 and all older families by adding two compatible
strings.

Signed-off-by: Abin Joseph <abin.joseph@amd.com>
Message-ID: <20250326065556.2053884-3-abin.joseph@amd.com>
State: pending