interrupt_controller: intc_xlnx: Fix HW designs with FAST interrupts enabled
When FAST interrupts are enabled, the interrupt entry point is determined by
the IVAR registers. The default value of the IVAR register is 0x10. With FAST
interrupts enabled in the AXI INTC core, when an interrupt occurs, the
processor's execution jumps to address 0x10, which may not be a valid entry
point, leading to a crash.

Since Zephyr does not support FAST interrupts, there is no way to configure
the IVAR registers with the correct FAST interrupt handler. To address this,
initialize the IVAR registers for all interrupt sources to point to the trap
handler base address, which serves as the entry point for all exceptions and
interrupts in the case of the MB-V.

Note:
- FAST interrupts are supported only when the AXI INTC core is paired with
  the MB-V processor, so the handling is specifically added for MB-V
- This change makes the behavior of FAST interrupts identical to that of
  normal interrupts. As a result, enabling FAST interrupts in the AXI INTC
  core will not lead to any improvement in interrupt latency.

Signed-off-by: Mubin Sayyed <mubin.sayyed@amd.com>
Tested-by: Appana Durga Kedareswara Rao <appana.durga.kedareswara.rao@amd.com>
Message-ID: <20250327115318.311332-1-mubin.sayyed@amd.com>
[v4.3.0 update
Origin: c8eb83d7b747 ("interrupt_controller: intc_xlnx: Fix HW designs with FAST interrupts enabled")
]
State: pending