drivers: dma: Add Xilinx AXI DMA driver
The Xilinx AXI DMA Controller is commonly used in FPGA designs.
For example, it is a part of the 1G/2.5G AXI Ethernet subsystem.
This patch adds a driver for the Xilinx AXI DMA that supports
single MM2S and S2MM channels as well as the control and status
streams used by the AXI Ethernet subsystem.

Signed-off-by: Eric Ackermann <eric.ackermann@cispa.de>
Signed-off-by: Venkatesh Odela <venkatesh.odela@amd.com>
Message-ID: <20250525165838.3067753-3-venkatesh.odela@amd.com>
State: upstream (c9ce311aaac2)