boards: amd: versalnet_rpu: Move OSPI node from SE7 board DTS to overlay
On the Tenzing2 SE7 board, both SPIPS and OSPI are enabled by default,
and the flash Ztest is executed for both drivers. However, when both the
driver nodes are defined within a same DTS file, the Ztest prioritizes
SPIPS over OSPI, resulting in the SPIPS driver being invoked exclusively
while the OSPI driver is skipped. To ensure the OSPI driver is tested,
remove OSPI node from SE7 board DTS and move it to an overlay under flash
Ztest.

Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@amd.com>
Message-ID: <20250910173404.3916211-5-amit.kumar-mahapatra@amd.com>