boards: amd: Add generic APU board support for Versal Gen 2 SoC (CortexA78)
Introduce a generic board configuration for the APU on the Versal Gen 2
SoC, which is based on the ARM Cortex-A78 processor.

This board setup provides a baseline environment for enabling and testing
Cortex-A78 features and peripheral integration on the APU subsystem.

Signed-off-by: Appana Durga Kedareswara rao <appana.durga.kedareswara.rao@amd.com>
State: upstream (3cdae0cf4b7c5b8c5892829001dedf29c7f48031)